+++ src/test/java/wooteco/subway/acceptance/PathAcceptanceTest.java
@@ -2,23 +2,23 @@ package wooteco.subway.acceptance;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
-import static wooteco.subway.common.TLine.LINE_NO_EXTRA_FARE;
-import static wooteco.subway.common.TLine.LINE_SIX;
-import static wooteco.subway.common.TLine.LINE_TWO;
-import static wooteco.subway.common.TStation.BOMUN;
-import static wooteco.subway.common.TStation.CHANGSIN;
-import static wooteco.subway.common.TStation.DONGMYO;
-import static wooteco.subway.common.TStation.SANGWANGSIMNI;
-import static wooteco.subway.common.TStation.SINDANG;
-import static wooteco.subway.common.TStation.WANGSIMNI;
-import static wooteco.subway.common.TestFixtures.CHILD_DEFAULT_FARE;
-import static wooteco.subway.common.TestFixtures.CHILD_MAX_AGE;
-import static wooteco.subway.common.TestFixtures.DEFAULT_AGE;
-import static wooteco.subway.common.TestFixtures.DEFAULT_FARE;
-import static wooteco.subway.common.TestFixtures.HALF_STANDARD_DISTANCE;
-import static wooteco.subway.common.TestFixtures.STANDARD_DISTANCE;
-import static wooteco.subway.common.TestFixtures.TEEN_DEFAULT_FARE;
-import static wooteco.subway.common.TestFixtures.TEEN_MAX_AGE;
+import static wooteco.subway.helper.TLine.LINE_NO_EXTRA_FARE;
+import static wooteco.subway.helper.TLine.LINE_SIX;
+import static wooteco.subway.helper.TLine.LINE_TWO;
+import static wooteco.subway.helper.TStation.BOMUN;
+import static wooteco.subway.helper.TStation.CHANGSIN;
+import static wooteco.subway.helper.TStation.DONGMYO;
+import static wooteco.subway.helper.TStation.SANGWANGSIMNI;
+import static wooteco.subway.helper.TStation.SINDANG;
+import static wooteco.subway.helper.TStation.WANGSIMNI;
+import static wooteco.subway.helper.TestFixtures.CHILD_DEFAULT_FARE;
+import static wooteco.subway.helper.TestFixtures.CHILD_MAX_AGE;
+import static wooteco.subway.helper.TestFixtures.DEFAULT_AGE;
+import static wooteco.subway.helper.TestFixtures.DEFAULT_FARE;
+import static wooteco.subway.helper.TestFixtures.HALF_STANDARD_DISTANCE;
+import static wooteco.subway.helper.TestFixtures.STANDARD_DISTANCE;
+import static wooteco.subway.helper.TestFixtures.TEEN_DEFAULT_FARE;
+import static wooteco.subway.helper.TestFixtures.TEEN_MAX_AGE;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
