+++ src/test/java/wooteco/subway/helper/LineAddAndRequest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.common;
+package wooteco.subway.helper;
 
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
