+++ src/test/java/wooteco/subway/helper/TestFixtures.java
@@ -1,4 +1,4 @@
-package wooteco.subway.common;
+package wooteco.subway.helper;
 
 import wooteco.subway.domain.Station;
