+++ src/test/java/wooteco/subway/repository/SectionRepositoryTest.java
@@ -2,11 +2,11 @@ package wooteco.subway.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
-import static wooteco.subway.common.TestFixtures.STANDARD_DISTANCE;
-import static wooteco.subway.common.TestFixtures.동묘앞역;
-import static wooteco.subway.common.TestFixtures.보문역;
-import static wooteco.subway.common.TestFixtures.신당역;
-import static wooteco.subway.common.TestFixtures.창신역;
+import static wooteco.subway.helper.TestFixtures.STANDARD_DISTANCE;
+import static wooteco.subway.helper.TestFixtures.동묘앞역;
+import static wooteco.subway.helper.TestFixtures.보문역;
+import static wooteco.subway.helper.TestFixtures.신당역;
+import static wooteco.subway.helper.TestFixtures.창신역;
 
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
