+++ src/test/java/wooteco/subway/service/LineServiceTest.java
@@ -4,12 +4,12 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.Assertions.tuple;
 import static org.junit.jupiter.api.Assertions.assertAll;
-import static wooteco.subway.common.TestFixtures.LINE_COLOR;
-import static wooteco.subway.common.TestFixtures.STANDARD_DISTANCE;
-import static wooteco.subway.common.TestFixtures.동묘앞역;
-import static wooteco.subway.common.TestFixtures.보문역;
-import static wooteco.subway.common.TestFixtures.신당역;
-import static wooteco.subway.common.TestFixtures.창신역;
+import static wooteco.subway.helper.TestFixtures.LINE_COLOR;
+import static wooteco.subway.helper.TestFixtures.STANDARD_DISTANCE;
+import static wooteco.subway.helper.TestFixtures.동묘앞역;
+import static wooteco.subway.helper.TestFixtures.보문역;
+import static wooteco.subway.helper.TestFixtures.신당역;
+import static wooteco.subway.helper.TestFixtures.창신역;
 
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
